# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserCreateResponseDto": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"accessToken": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"accessToken",
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserCreateDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"nickname": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"year",
					"nickname",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserLoginResponseDto": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"accessToken": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"accessToken",
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserLoginDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PostBaseResponseDto": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RecipeResponse": {
				"properties": {
					"imageURL": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"writerInfo": {
						"type": "string"
					}
				},
				"required": [
					"imageURL",
					"title",
					"writerInfo"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TotalRecipeResponseDto": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/RecipeResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RecipeResponseDto": {
				"properties": {
					"imageURL": {
						"type": "string"
					},
					"writer": {
						"type": "string"
					},
					"food": {
						"type": "string"
					},
					"ingredient": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"imageURL",
					"writer",
					"food",
					"ingredient",
					"content"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EachRecipeResponseDto": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/RecipeResponseDto"
					}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "Bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "ccst-server",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"email": "jokj624@gmail.com",
			"name": "채정아"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/signup": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreateResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateDto"
							}
						}
					}
				}
			}
		},
		"/auth/signin": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLoginResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Not Found - 이메일에 해당하는 사용자 정보 없음"
					},
					"409": {
						"description": "Invalid - 비밀번호가 올바르지 않음"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLoginDto"
							}
						}
					}
				}
			}
		},
		"/recipe": {
			"post": {
				"operationId": "CreateRecipe",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostBaseResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Not Found - 이름에 해당하는 사용자 정보 없음"
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									},
									"food": {
										"type": "string"
									},
									"content": {
										"type": "string"
									},
									"ingredient": {
										"type": "string"
									}
								},
								"required": [
									"file",
									"food",
									"content",
									"ingredient"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetRecipe",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TotalRecipeResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [],
				"parameters": []
			}
		},
		"/recipe/{recipeId}": {
			"get": {
				"operationId": "GetRecipeById",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EachRecipeResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}